+++ script/tool/test/pubspec_check_command_test.dart
@@ -157,6 +157,21 @@ ${_devDependenciesSection()}
 ${_falseSecretsSection()}
 ''');
 
+      pluginDirectory
+          .childDirectory('example')
+          .childFile('pubspec.yaml')
+          .writeAsStringSync('''
+${_headerSection(
+        'plugin_example',
+        publishable: false,
+        includeRepository: false,
+        includeIssueTracker: false,
+      )}
+${_environmentSection()}
+${_dependenciesSection()}
+${_flutterSection()}
+''');
+
       final List<String> output = await runCapturingPrint(runner, <String>[
         'pubspec-check',
       ]);
@@ -172,15 +187,31 @@ ${_falseSecretsSection()}
     });
 
     test('passes for a Flutter package following conventions', () async {
-      final Directory pluginDirectory = createFakePlugin('plugin', packagesDir);
+      final Directory packageDirectory =
+          createFakePackage('a_package', packagesDir);
 
-      pluginDirectory.childFile('pubspec.yaml').writeAsStringSync('''
-${_headerSection('plugin')}
+      packageDirectory.childFile('pubspec.yaml').writeAsStringSync('''
+${_headerSection('a_package')}
 ${_environmentSection()}
 ${_dependenciesSection()}
 ${_devDependenciesSection()}
 ${_flutterSection()}
 ${_falseSecretsSection()}
+''');
+
+      packageDirectory
+          .childDirectory('example')
+          .childFile('pubspec.yaml')
+          .writeAsStringSync('''
+${_headerSection(
+        'a_package',
+        publishable: false,
+        includeRepository: false,
+        includeIssueTracker: false,
+      )}
+${_environmentSection()}
+${_dependenciesSection()}
+${_flutterSection()}
 ''');
 
       final List<String> output = await runCapturingPrint(runner, <String>[
@@ -190,8 +221,8 @@ ${_falseSecretsSection()}
       expect(
         output,
         containsAllInOrder(<Matcher>[
-          contains('Running for plugin...'),
-          contains('Running for plugin/example...'),
+          contains('Running for a_package...'),
+          contains('Running for a_package/example...'),
           contains('No issues found!'),
         ]),
       );
@@ -221,7 +252,8 @@ ${_dependenciesSection()}
     });
 
     test('fails when homepage is included', () async {
-      final Directory pluginDirectory = createFakePlugin('plugin', packagesDir);
+      final Directory pluginDirectory =
+          createFakePlugin('plugin', packagesDir, examples: <String>[]);
 
       pluginDirectory.childFile('pubspec.yaml').writeAsStringSync('''
 ${_headerSection('plugin', isPlugin: true, includeHomepage: true)}
@@ -248,7 +280,8 @@ ${_devDependenciesSection()}
     });
 
     test('fails when repository is missing', () async {
-      final Directory pluginDirectory = createFakePlugin('plugin', packagesDir);
+      final Directory pluginDirectory =
+          createFakePlugin('plugin', packagesDir, examples: <String>[]);
 
       pluginDirectory.childFile('pubspec.yaml').writeAsStringSync('''
 ${_headerSection('plugin', isPlugin: true, includeRepository: false)}
@@ -274,7 +307,8 @@ ${_devDependenciesSection()}
     });
 
     test('fails when homepage is given instead of repository', () async {
-      final Directory pluginDirectory = createFakePlugin('plugin', packagesDir);
+      final Directory pluginDirectory =
+          createFakePlugin('plugin', packagesDir, examples: <String>[]);
 
       pluginDirectory.childFile('pubspec.yaml').writeAsStringSync('''
 ${_headerSection('plugin', isPlugin: true, includeHomepage: true, includeRepository: false)}
@@ -301,7 +335,8 @@ ${_devDependenciesSection()}
     });
 
     test('fails when repository is incorrect', () async {
-      final Directory pluginDirectory = createFakePlugin('plugin', packagesDir);
+      final Directory pluginDirectory =
+          createFakePlugin('plugin', packagesDir, examples: <String>[]);
 
       pluginDirectory.childFile('pubspec.yaml').writeAsStringSync('''
 ${_headerSection('plugin', isPlugin: true, repositoryPackagesDirRelativePath: 'different_plugin')}
@@ -327,7 +362,8 @@ ${_devDependenciesSection()}
     });
 
     test('fails when issue tracker is missing', () async {
-      final Directory pluginDirectory = createFakePlugin('plugin', packagesDir);
+      final Directory pluginDirectory =
+          createFakePlugin('plugin', packagesDir, examples: <String>[]);
 
       pluginDirectory.childFile('pubspec.yaml').writeAsStringSync('''
 ${_headerSection('plugin', isPlugin: true, includeIssueTracker: false)}
@@ -353,8 +389,9 @@ ${_devDependenciesSection()}
     });
 
     test('fails when description is too short', () async {
-      final Directory pluginDirectory =
-          createFakePlugin('a_plugin', packagesDir.childDirectory('a_plugin'));
+      final Directory pluginDirectory = createFakePlugin(
+          'a_plugin', packagesDir.childDirectory('a_plugin'),
+          examples: <String>[]);
 
       pluginDirectory.childFile('pubspec.yaml').writeAsStringSync('''
 ${_headerSection('plugin', isPlugin: true, description: 'Too short')}
@@ -383,7 +420,8 @@ ${_devDependenciesSection()}
     test(
         'allows short descriptions for non-app-facing parts of federated plugins',
         () async {
-      final Directory pluginDirectory = createFakePlugin('plugin', packagesDir);
+      final Directory pluginDirectory =
+          createFakePlugin('plugin', packagesDir, examples: <String>[]);
 
       pluginDirectory.childFile('pubspec.yaml').writeAsStringSync('''
 ${_headerSection('plugin', isPlugin: true, description: 'Too short')}
@@ -410,7 +448,8 @@ ${_devDependenciesSection()}
     });
 
     test('fails when description is too long', () async {
-      final Directory pluginDirectory = createFakePlugin('plugin', packagesDir);
+      final Directory pluginDirectory =
+          createFakePlugin('plugin', packagesDir, examples: <String>[]);
 
       const String description = 'This description is too long. It just goes '
           'on and on and on and on and on. pub.dev will down-score it because '
@@ -442,7 +481,8 @@ ${_devDependenciesSection()}
     });
 
     test('fails when environment section is out of order', () async {
-      final Directory pluginDirectory = createFakePlugin('plugin', packagesDir);
+      final Directory pluginDirectory =
+          createFakePlugin('plugin', packagesDir, examples: <String>[]);
 
       pluginDirectory.childFile('pubspec.yaml').writeAsStringSync('''
 ${_headerSection('plugin', isPlugin: true)}
@@ -469,7 +509,8 @@ ${_environmentSection()}
     });
 
     test('fails when flutter section is out of order', () async {
-      final Directory pluginDirectory = createFakePlugin('plugin', packagesDir);
+      final Directory pluginDirectory =
+          createFakePlugin('plugin', packagesDir, examples: <String>[]);
 
       pluginDirectory.childFile('pubspec.yaml').writeAsStringSync('''
 ${_headerSection('plugin', isPlugin: true)}
@@ -496,7 +537,8 @@ ${_devDependenciesSection()}
     });
 
     test('fails when dependencies section is out of order', () async {
-      final Directory pluginDirectory = createFakePlugin('plugin', packagesDir);
+      final Directory pluginDirectory =
+          createFakePlugin('plugin', packagesDir, examples: <String>[]);
 
       pluginDirectory.childFile('pubspec.yaml').writeAsStringSync('''
 ${_headerSection('plugin', isPlugin: true)}
@@ -550,7 +592,8 @@ ${_dependenciesSection()}
     });
 
     test('fails when false_secrets section is out of order', () async {
-      final Directory pluginDirectory = createFakePlugin('plugin', packagesDir);
+      final Directory pluginDirectory =
+          createFakePlugin('plugin', packagesDir, examples: <String>[]);
 
       pluginDirectory.childFile('pubspec.yaml').writeAsStringSync('''
 ${_headerSection('plugin', isPlugin: true)}
@@ -580,7 +623,8 @@ ${_devDependenciesSection()}
     test('fails when an implemenation package is missing "implements"',
         () async {
       final Directory pluginDirectory = createFakePlugin(
-          'plugin_a_foo', packagesDir.childDirectory('plugin_a'));
+          'plugin_a_foo', packagesDir.childDirectory('plugin_a'),
+          examples: <String>[]);
 
       pluginDirectory.childFile('pubspec.yaml').writeAsStringSync('''
 ${_headerSection('plugin_a_foo', isPlugin: true)}
@@ -608,7 +652,8 @@ ${_devDependenciesSection()}
     test('fails when an implemenation package has the wrong "implements"',
         () async {
       final Directory pluginDirectory = createFakePlugin(
-          'plugin_a_foo', packagesDir.childDirectory('plugin_a'));
+          'plugin_a_foo', packagesDir.childDirectory('plugin_a'),
+          examples: <String>[]);
 
       pluginDirectory.childFile('pubspec.yaml').writeAsStringSync('''
 ${_headerSection('plugin_a_foo', isPlugin: true)}
@@ -636,7 +681,8 @@ ${_devDependenciesSection()}
 
     test('passes for a correct implemenation package', () async {
       final Directory pluginDirectory = createFakePlugin(
-          'plugin_a_foo', packagesDir.childDirectory('plugin_a'));
+          'plugin_a_foo', packagesDir.childDirectory('plugin_a'),
+          examples: <String>[]);
 
       pluginDirectory.childFile('pubspec.yaml').writeAsStringSync('''
 ${_headerSection(
@@ -663,8 +709,9 @@ ${_devDependenciesSection()}
     });
 
     test('fails when a "default_package" looks incorrect', () async {
-      final Directory pluginDirectory =
-          createFakePlugin('plugin_a', packagesDir.childDirectory('plugin_a'));
+      final Directory pluginDirectory = createFakePlugin(
+          'plugin_a', packagesDir.childDirectory('plugin_a'),
+          examples: <String>[]);
 
       pluginDirectory.childFile('pubspec.yaml').writeAsStringSync('''
 ${_headerSection(
@@ -702,8 +749,9 @@ ${_devDependenciesSection()}
     test(
         'fails when a "default_package" does not have a corresponding dependency',
         () async {
-      final Directory pluginDirectory =
-          createFakePlugin('plugin_a', packagesDir.childDirectory('plugin_a'));
+      final Directory pluginDirectory = createFakePlugin(
+          'plugin_a', packagesDir.childDirectory('plugin_a'),
+          examples: <String>[]);
 
       pluginDirectory.childFile('pubspec.yaml').writeAsStringSync('''
 ${_headerSection(
@@ -739,8 +787,9 @@ ${_devDependenciesSection()}
     });
 
     test('passes for an app-facing package without "implements"', () async {
-      final Directory pluginDirectory =
-          createFakePlugin('plugin_a', packagesDir.childDirectory('plugin_a'));
+      final Directory pluginDirectory = createFakePlugin(
+          'plugin_a', packagesDir.childDirectory('plugin_a'),
+          examples: <String>[]);
 
       pluginDirectory.childFile('pubspec.yaml').writeAsStringSync('''
 ${_headerSection(
@@ -769,8 +818,8 @@ ${_devDependenciesSection()}
     test('passes for a platform interface package without "implements"',
         () async {
       final Directory pluginDirectory = createFakePlugin(
-          'plugin_a_platform_interface',
-          packagesDir.childDirectory('plugin_a'));
+          'plugin_a_platform_interface', packagesDir.childDirectory('plugin_a'),
+          examples: <String>[]);
 
       pluginDirectory.childFile('pubspec.yaml').writeAsStringSync('''
 ${_headerSection(
@@ -799,7 +848,8 @@ ${_devDependenciesSection()}
 
     test('validates some properties even for unpublished packages', () async {
       final Directory pluginDirectory = createFakePlugin(
-          'plugin_a_foo', packagesDir.childDirectory('plugin_a'));
+          'plugin_a_foo', packagesDir.childDirectory('plugin_a'),
+          examples: <String>[]);
 
       // Environment section is in the wrong location.
       // Missing 'implements'.
@@ -829,7 +879,8 @@ ${_environmentSection()}
     });
 
     test('ignores some checks for unpublished packages', () async {
-      final Directory pluginDirectory = createFakePlugin('plugin', packagesDir);
+      final Directory pluginDirectory =
+          createFakePlugin('plugin', packagesDir, examples: <String>[]);
 
       // Missing metadata that is only useful for published packages, such as
       // repository and issue tracker.
@@ -886,7 +937,7 @@ ${_devDependenciesSection()}
 
     test('repository check works', () async {
       final Directory packageDirectory =
-          createFakePackage('package', packagesDir);
+          createFakePackage('package', packagesDir, examples: <String>[]);
 
       packageDirectory.childFile('pubspec.yaml').writeAsStringSync('''
 ${_headerSection('package')}