+++ script/tool/CHANGELOG.md
@@ -1,3 +1,7 @@
+## NEXT
+
+- `drive-examples` now supports non-plugin packages.
+
 ## 0.8.4
 
 - `readme-check` now validates that there's a info tag on code blocks to